+++ src/Record/GroupByQueryInterface.php
@@ -6,4 +6,5 @@ use Pyncer\Database\Record\RecordQueryInterface;
 interface GroupByQueryInterface extends RecordQueryInterface
 {
     public function groupBy(string|array ...$columns): static;
+    public function hasGroupBy(): bool;
 }